requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-keyvault-secrets==4.7.0
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
colorama==0.4.6
crashtest==0.4.1
cryptography==40.0.2
dataclasses-json==0.5.7
distlib==0.3.6
Django==4.2
dulwich==0.21.3
filelock==3.11.0
frozenlist==1.3.3
gptcache==0.1.20
greenlet==2.0.2
html5lib==1.1
idna==3.4
importlib-metadata==6.4.1
installer==0.7.0
isodate==0.6.1
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
langchain==0.0.142
llama-index==0.5.25
lockfile==0.12.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
msal==1.22.0
msal-extensions==1.0.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.0
portalocker==2.7.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
PyPDF2==3.0.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
rapidfuzz==2.15.1
regex==2023.3.23
requests==2.28.2
requests-toolbelt==0.10.1
shellingham==1.5.0.post1
six==1.16.0
SQLAlchemy==1.4.47
sqlparse==0.4.4
tenacity==8.2.2
tiktoken==0.3.3
tomli==2.0.1
tomlkit==0.11.7
tqdm==4.65.0
trove-classifiers==2023.3.9
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
webencodings==0.5.1
yarl==1.9.2
zipp==3.15.0
